[ALPS03846483] disp:[coverity] integer overflow(4260135)
[Detail]
integer overflow

MTK-Commit-Id: 72a8011be317c9538d6216d79dc19dc501c4f3e2

CR-Id: ALPS03846483
Feature: Display Driver (DSI Interface)
(cherry picked from commit 6d3d685d7946e8a600774a802616261e543cb26f)

+++ drivers/misc/mediatek/video/mt6765/dispsys/ddp_ovl.c
@@ -1269,7 +1269,7 @@ static unsigned long long full_trans_bw_calc(struct sbch *data,
 	/* caculate the layer bw */
 	if (data->full_trans_en) {
 		bpp = ufmt_get_Bpp(cfg->fmt);
-		bw_sum = cfg->dst_w * cfg->dst_h * bpp;
+		bw_sum = (unsigned long long)cfg->dst_w * cfg->dst_h * bpp;
 	}
 
 	return bw_sum;
